+++ 0pointers/src/lib.rs
@@ -9,25 +9,25 @@ use std::mem::transmute;
 ///
 /// Dereferencing is preferred over transmute here since it's generally slightly more restrictive.
 #[no_mangle]
-pub extern fn createCounter(val: usize) -> *mut Counter {
+pub extern fn createCounter(val: u32) -> *mut Counter {
     let _counter = &mut *Box::new(Counter::new(val));
     _counter
 }
 
 #[no_mangle]
-pub extern fn getCounterValue(ptr: *mut Counter) -> usize {
+pub extern fn getCounterValue(ptr: *mut Counter) -> u32 {
     let mut _counter = unsafe { &mut *ptr };
     _counter.get()
 }
 
 #[no_mangle]
-pub extern fn incrementCounterBy(ptr: *mut Counter, by: usize) -> usize {
+pub extern fn incrementCounterBy(ptr: *mut Counter, by: u32) -> u32 {
     let mut _counter = unsafe { &mut *ptr };
     _counter.incr(by)
 }
 
 #[no_mangle]
-pub extern fn decrementCounterBy(ptr: *mut Counter, by: usize) -> usize {
+pub extern fn decrementCounterBy(ptr: *mut Counter, by: u32) -> u32 {
     let mut _counter = unsafe { &mut *ptr };
     _counter.decr(by)
 }
